test(e2e): retire docker-unreachable gateway script
